python/requirements.txt
bme680==2.0.0
mysql-connector-python==8.2.0
protobuf==4.21.12
RPi.GPIO==0.7.1
smbus2==0.5.0
